CSCTTV-4139 Fix error in profile editor data. Use long instead of int for ProfileEditorItemMeta.Id

+++ aspnetcore/src/api/Models/ProfileEditor/Items/ProfileEditorItemMeta.cs
@@ -2,7 +2,7 @@
 {
     public partial class ProfileEditorItemMeta
     {
-        public ProfileEditorItemMeta(int id, int type, bool? show, bool? primaryValue)
+        public ProfileEditorItemMeta(long id, int type, bool? show, bool? primaryValue)
         {
             Id = id;
             Type = type;
@@ -11,7 +11,7 @@
             TemporaryUniqueId = ulong.Parse(Type.ToString() + Id.ToString());
         }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int Type { get; set; }
         public bool? Show { get; set; }
         public bool? PrimaryValue { get; set; }
